playlist handler silently ignores Watch Later ID & falsy (empty) IDs
Ignoring None, 0, '', etc. is for saving on useless API requests, and
for avoiding exceptions from trying to call `.upper()` on a non-string.

Ignoring the WL playlist is for saving on useless requests too, since
YouTube's API won't return any data about it unless there's an
authenticated user associated with the request (since Watch Later is
stored on a per-user basis).