Use `BooleanAttribute` setting type (requires Sopel 7.1+)
The person responsible for implementing this new setting type (also me)
probably should have set it up to accept the old `ValidatedAttribute`
style silently in 7.x, and start warning in 8.0, but oh well. Even if
that's changed for an upcoming patch release of Sopel, this patch still
needs to be written at some point.