youtube: stop trying to fetch/process dislike count
The YouTube API officially stopped returning public dislike info as of
today, 13 December 2021.

It's possible to apply for an exemption and continue to receive dislike
data on unauthenticated API calls, but only on the condition that said
data is not displayed to end-users. Since showing info to end-users is
the whole point of this plugin, we won't bother asking.

https://support.google.com/youtube/thread/134791097/update-to-youtube-dislike-counts