Safer handling of live stream info
Viewer count isn't always available for live videos, whether it's hidden
by the content owner or just not supported for the particular type of
live stream in use.

While we're at it, let's also make sure that `in` will always work for
the `live_info` shortcut variable, meaning it has to be an empty dict
instead of `None` if not present in the API result.