build, init: move version number to setup.py
No longer try/catch and hide import errors. This confused someone on IRC
(and me, too, when I tried to help troubleshoot).

Not crazy about the runtime dependency on `pkg_resources` (meaning
`setuptools` is needed *after* build/install), but this is what Sopel
itself does in 7.x (8.x will move to `importlib_metadata`, as indicated
by the relevant comment in this patch).