Add very basic error logging
The idea here is not to be comprehensive, but to make it more obvious
when a Google Cloud project is misconfigured by passing the exact reason
string returned by the API client library in its HTTPError exception.

This string gives clear instructions for fixing common issues, like
forgetting to enable the YouTube Data API for the project that issued
the bot's API key.